\name{Rindsel-package}
\Rdversion{1.0}
\alias{Rindsel-package}
\alias{Rindsel-package}
\docType{package}
\title{R package for Selection Indices}

\description{
Calculated some different indices.
}

\details{
\tabular{ll}{
Package:  \tab Rindsel\cr
Type:     \tab Package\cr
Version:  \tab 5.1\cr
Date:     \tab 28.02.2023\cr
License:  \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}


The following functions can be used for the estimation of Selection Index(SI):

\tabular{ll}{
\link{BLPSI}:	   \tab Base Linear Phenotypic Selection Index.\cr
\link{SmithIndex}: \tab Linear Phenotypic Selection Index.\cr
\link{ESIMIndex}:  \tab Eigen Selection Index Method.\cr
\link{PPGLPSI}:	   \tab Predeterminated Proportional Gain Linear Phenotypic Selection Index.\cr
\link{KNIndex}:    \tab Restrictive Linear Phenotypic Selection Index.\cr
\link{RESIMIndex}: \tab Restrictive Eigen Selection Index Method.\cr
\link{LTIndex}:    \tab Linear Marker Selection Index.\cr
\link{MESIMIndex}: \tab Molecular Eigen Selection Index Method.\cr
\link{LGSI}:	   \tab Linear Genomic Selection Index.\cr
}



